Set up bindings for an empty HandleGraph implementation
