--- src/bindings.cpp
@@ -1,14 +1,0 @@
-#include "handlegraph/handle_graph.hpp"
-
-int fact(int n) {
-    if (n < 0){ /* This should probably return an error, but this is simpler */
-        return 0;
-    }
-    if (n == 0) {
-        return 1;
-    }
-    else {
-        /* testing for overflow would be a good idea here */
-        return n * fact(n-1);
-    }
-}
