+++ src/include/handlegraph/handle_graph.hpp
@@ -11,8 +11,6 @@
 #include <functional>
 #include <string>
 
-int fact(int n);
-
 namespace handlegraph {
